+++ parsers/src/main/java/module-info.java
@@ -24,6 +24,7 @@ module org.chocosolver.parsers {
 
     opens org.chocosolver.parser to args4j, testng;
     opens org.chocosolver.parser.mps to args4j;
-    opens org.chocosolver.parser.xcsp to args4j;
     opens org.chocosolver.parser.flatzinc to args4j;
+    opens org.chocosolver.parser.xcsp to args4j;
+    opens org.chocosolver.parser.dimacs to args4j;
 }