+++ solver/src/main/java/org/chocosolver/solver/constraints/nary/nvalue/amnv/mis/Rk.java
@@ -87,7 +87,7 @@ public class Rk implements F {
             }
             inMIS.set(idx);
             out.set(idx);
-            ISetIterator nei = graph.getNeighOf(idx).iterator();
+            ISetIterator nei = graph.getNeighborsOf(idx).iterator();
             while (nei.hasNext()) {
                 out.set(nei.nextInt());
             }