+++ solver/src/main/java/org/chocosolver/solver/variables/view/IView.java
@@ -11,14 +11,9 @@ package org.chocosolver.solver.variables.view;
 
 import org.chocosolver.solver.ICause;
 import org.chocosolver.solver.exception.ContradictionException;
-import org.chocosolver.solver.learn.ExplanationForSignedClause;
-import org.chocosolver.solver.variables.IntVar;
 import org.chocosolver.solver.variables.Variable;
 import org.chocosolver.solver.variables.events.IEventType;
 import org.chocosolver.solver.variables.events.IntEventType;
-import org.chocosolver.util.objects.setDataStructures.iterable.IntIterableRangeSet;
-
-import static org.chocosolver.util.objects.setDataStructures.iterable.IntIterableSetUtils.unionOf;
 
 /**
  * An interface to define views.
@@ -28,17 +23,36 @@ import static org.chocosolver.util.objects.setDataStructures.iterable.IntIterabl
  * This is intend to replace very specific propagator such as equality.
  * <br/>
  *
+ * This is an implementation of domain views, as described in:
+ * Van Hentenryck P., Michel L. (2014) Domain Views for Constraint Programming
+ * https://link.springer.com/chapter/10.1007/978-3-319-10428-7_51
+ *
+ * --- UPDATE 08/03/2021 (Dimitri Justeau-Allaire) ---
+ * Views are now by default defined as global for more genericity.
+ * Global views are an extension of views for observing several variables.
+ * A global view is a variable that does not declare any domain but relies on an array of variables.
+ * Global views are particularly useful to provide a more efficient and more expressive way than channelling
+ * constraints to link different types of variables such as a SetVar and an array of IntVar.
+ * ---------------------------------------------------
+ *
  * @author Charles Prud'homme
  * @since 26/08/11
  */
-public interface IView extends ICause, Variable {
+public interface IView<V extends Variable> extends ICause, Variable {
 
     /**
-     * Return the basis variable
+     * Return the array of observed variables
      *
-     * @return variable observed
+     * @return variables observed
      */
-    IntVar getVariable();
+    V[] getVariables();
+
+    /**
+     * @return The number of observed variables
+     */
+    default int getNbObservedVariables() {
+        return getVariables().length;
+    }
 
     /**
      * This methods is related to explanations, it binds an event occurring on the observed
@@ -53,17 +67,9 @@ public interface IView extends ICause, Variable {
     /**
      * To notify a view that the variable is observed has been modified.
      * @param event the event received by the observed variable
+     * @param variableIdx the index of the variable in the view's observed variables
      * @throws ContradictionException if a failure occurs
      */
-    void notify(IEventType event) throws ContradictionException;
+    void notify(IEventType event, int variableIdx) throws ContradictionException;
 
-    default void explain(int p, ExplanationForSignedClause explanation) {
-        IntVar pivot = explanation.readVar(p);
-        IntVar other = (this == pivot ? getVariable() : (IntVar)this);
-        IntIterableRangeSet dom = explanation.complement(other);
-        other.unionLit(dom, explanation);
-        dom = explanation.complement(pivot);
-        unionOf(dom, explanation.readDom(p));
-        pivot.intersectLit(dom, explanation);
-    }
 }